Code revised
